(mostly) fix markers, restructure API
