+++ src/main/java/gregtech/api/unification/material/info/MaterialFlags.java
@@ -1,4 +1,4 @@
-package gregtech.api.unification.material.type;
+package gregtech.api.unification.material.info;
 
 import gregtech.api.unification.material.Material;
