+++ src/main/java/gregtech/api/unification/material/materials/SecondDegreeMaterials.java
@@ -3,9 +3,9 @@ package gregtech.api.unification.material.materials;
 import gregtech.api.GTValues;
 import gregtech.api.unification.material.Material;
 
-import static gregtech.api.unification.material.MaterialIconSet.*;
+import static gregtech.api.unification.material.info.MaterialIconSet.*;
 import static gregtech.api.unification.material.Materials.*;
-import static gregtech.api.unification.material.type.MaterialFlags.*;
+import static gregtech.api.unification.material.info.MaterialFlags.*;
 
 public class SecondDegreeMaterials {
