+++ src/main/java/gregtech/common/blocks/modelfactories/BlockCompressedFactory.java
@@ -4,8 +4,8 @@ import com.google.common.base.Joiner;
 import com.google.common.collect.ImmutableList;
 import gregtech.api.model.AbstractBlockModelFactory;
 import gregtech.api.model.ResourcePackHook;
+import gregtech.api.unification.material.info.MaterialIconType;
 import gregtech.api.unification.material.Material;
-import gregtech.api.unification.material.MaterialIconType;
 import gregtech.common.blocks.BlockCompressed;
 import net.minecraft.block.Block;
 import net.minecraftforge.fml.relauncher.Side;
