+++ src/main/java/gregtech/common/blocks/modelfactories/BlockFrameFactory.java
@@ -2,7 +2,7 @@ package gregtech.common.blocks.modelfactories;
 
 import gregtech.api.model.AbstractBlockModelFactory;
 import gregtech.api.model.ResourcePackHook;
-import gregtech.api.unification.material.MaterialIconType;
+import gregtech.api.unification.material.info.MaterialIconType;
 import gregtech.api.unification.material.Material;
 import gregtech.common.blocks.BlockFrame;
 import net.minecraft.block.Block;
